streaming-kafka/pyproject.toml
[project]
name = "streaming-kafka"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python>=2.2.15",
    "requests-sse>=0.5.2",
    "ruff>=0.14.1",
]
